requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
GPUtil==1.4.0
idna==3.4
pip==22.2.2
psutil==5.9.3
python-telegram-bot==13.14
pytz-deprecation-shim==0.1.0.post0
pytz==2022.5
requests==2.28.1
setuptools==65.3.0
six==1.16.0
speedtest-cli==2.1.3
termcolor==2.0.1
tornado==6.1
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
wheel==0.37.1
